requirements.txt
cached-property==1.4.0
Django==2.0.3
django-filter==1.1.0
django-filters==0.2.1
django-url-filter==0.3.5
djangorestframework==3.7.6
enum-compat==0.0.2
Markdown==2.6.11
pytz==2018.3
six==1.11.0
